# api/docs/swagger.yaml
definitions:
  error_handler.ErrorResponse:
    properties:
      error:
        type: string
      error_detail:
        items:
          type: string
        type: array
      status:
        type: integer
    type: object
  request.CreateUserRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      prefecture:
        type: integer
      sex:
        enum:
        - male
        - female
        - other
        type: string
    required:
    - birthday
    - email
    - name
    - password
    - prefecture
    - sex
    type: object
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.SendLikeRequest:
    properties:
      receiver_id:
        description: いいねを受け取るユーザー
        type: string
    required:
    - receiver_id
    type: object
  request.SendMessageRequest:
    properties:
      message:
        description: メッセージ
        type: string
      receiver_id:
        description: メッセージを受け取るユーザー
        type: string
    required:
    - message
    - receiver_id
    type: object
  response.AuthenticationResponse:
    properties:
      api_token:
        description: IDトークン
        type: string
    required:
    - api_token
    type: object
  response.InitialResponse:
    properties:
      current_version:
        description: current version
        type: string
      min_version:
        description: min version
        type: string
    required:
    - current_version
    - min_version
    type: object
  response.LoginUserResponse:
    properties:
      authentication:
        $ref: '#/definitions/response.AuthenticationResponse'
      me:
        $ref: '#/definitions/response.MeResponse'
    required:
    - authentication
    - me
    type: object
  response.MeResponse:
    properties:
      avatar:
        description: アバター
        type: string
      birthday:
        description: 生年月日
        type: string
      email:
        description: メールアドレス
        type: string
      id:
        description: ID
        type: string
      like:
        description: いいね数
        type: integer
      message:
        description: メッセージ
        type: string
      name:
        description: 名前
        type: string
      prefecture:
        $ref: '#/definitions/response.PrefectureResponse'
        description: 都道府県
      sex:
        description: 性別
        enum:
        - male
        - female
        - other
        type: string
    required:
    - birthday
    - email
    - id
    - like
    - name
    - prefecture
    - sex
    type: object
  response.PrefectureResponse:
    properties:
      code:
        description: 都道府県コード
        type: integer
      name:
        description: 都道府県名
        type: string
    required:
    - code
    - name
    type: object
  response.ReceiveMessageResponse:
    properties:
      message:
        description: メッセージ
        type: string
      receive_at:
        description: 受信日時
        type: string
      sender:
        $ref: '#/definitions/response.UserResponse'
        description: 送り手
    required:
    - message
    - receive_at
    - sender
    type: object
  response.SendLikeResponse:
    properties:
      receiver:
        $ref: '#/definitions/response.UserResponse'
        description: 受け取り手
      sent_at:
        description: 送信日時
        type: string
    required:
    - receiver
    - sent_at
    type: object
  response.UserResponse:
    properties:
      avatar:
        description: アバター
        type: string
      birthday:
        description: 生年月日
        type: string
      id:
        description: ID
        type: string
      like:
        description: いいね数
        type: integer
      message:
        description: メッセージ
        type: string
      name:
        description: 名前
        type: string
      prefecture:
        $ref: '#/definitions/response.PrefectureResponse'
        description: 都道府県
      sex:
        description: 性別
        enum:
        - male
        - female
        - other
        type: string
    required:
    - birthday
    - id
    - like
    - name
    - prefecture
    - sex
    type: object
host: localhost:8080
info:
  contact: {}
  description: Matching app
  title: Matching app
  version: "1.0"
paths:
  /api/v1/health-check:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: health check
      tags:
      - healthCheck
  /api/v1/initial:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.InitialResponse'
      summary: アプリ起動時にコールする
      tags:
      - initial
  /api/v1/likes:
    post:
      consumes:
      - application/json
      description: いいねを送る
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: いいねを送る情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SendLikeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: いいねを送る
      tags:
      - like
  /api/v1/likes/{receiverID}/cancel:
    delete:
      consumes:
      - application/json
      description: いいねを取り消す
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 取り消しにするユーザーID
        in: path
        name: receiverID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: いいねを取り消す
      tags:
      - like
  /api/v1/likes/receive:
    get:
      consumes:
      - application/json
      description: 登録日が新しいもの順で返す
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.SendLikeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: 受信したいいね一覧を取得する
      tags:
      - like
  /api/v1/likes/send:
    get:
      consumes:
      - application/json
      description: 登録日が新しいもの順で返す
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.SendLikeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: 送信したいいね一覧を取得する
      tags:
      - like
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: ログイン処理を行う。JWTを新たに発行する
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: ログイン
      tags:
      - auth
  /api/v1/messages:
    get:
      consumes:
      - application/json
      description: 送信者別の受信メッセージ(最新の1通のみ)を受信日が最新のもの順に返す。
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ReceiveMessageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: 受信済みメッセージを取得
      tags:
      - message
    post:
      consumes:
      - application/json
      description: メッセージを送る
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: メッセージを送る情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SendMessageRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: メッセージを送る
      tags:
      - message
  /api/v1/messages/{id}/read:
    put:
      consumes:
      - application/json
      description: メッセージを既読にする
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 既読にするメッセージID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: メッセージを既読にする
      tags:
      - message
  /api/v1/prefectures:
    get:
      consumes:
      - application/json
      description: 都道府県のコードと名前のリストを取得する
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.PrefectureResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: 都道府県リスト取得
      tags:
      - prefecture
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: 50件ずつ取得。検索条件がない場合は登録日が新しい順に返す（今の所）
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: ページ
        in: query
        name: page
        type: integer
      - description: 都道府県
        in: query
        name: prefecture
        type: integer
      - description: 年齢 From
        in: query
        name: from_age
        type: integer
      - description: 年齢 To
        in: query
        name: to_age
        type: integer
      - description: 並び順
        in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: ユーザー情報全件取得
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 新しいユーザーを作成する
      parameters:
      - description: ユーザー情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: ユーザー新規作成
      tags:
      - user
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: 指定したユーザーの詳細情報を取得する
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: ユーザー詳細情報取得
      tags:
      - user
  /api/v1/users/info/me:
    get:
      consumes:
      - application/json
      description: ログイン中のユーザー情報を取得する
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: ログインユーザー情報取得
      tags:
      - user
  /api/v1/users/pickup/today:
    get:
      consumes:
      - application/json
      description: ログインユーザーとは異なる性別のユーザーを20件返す
      parameters:
      - description: ログイン時に取得したIDトークン(Bearer)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: 本日のピックアップユーザー取得
      tags:
      - user
swagger: "2.0"